45_Flight_Deal_GoogleForm/requirements.txt
python-dotenv==1.1.1
requests==2.32.5
twilio==9.8.4
